Extract hud and info_screens
